requirements.txt
chromadb==0.5.0
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
openai==1.33.0
python-dotenv==1.0.1
tiktoken==0.7.0]
groq==0.9.0
langchain-groq==0.1.5
